Split out use of httpsProxy preference
