+++ src/common/fetch/fetch.injectable.ts
@@ -3,32 +3,17 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { getInjectable } from "@ogre-tools/injectable";
-import { HttpsProxyAgent } from "hpagent";
-import type * as FetchModule from "node-fetch";
-import userStoreInjectable from "../user-store/user-store.injectable";
-
-const { NodeFetch: { default: fetch }} = require("../../../build/webpack/node-fetch.bundle") as { NodeFetch: typeof FetchModule };
-
-type Response = FetchModule.Response;
-type RequestInit = FetchModule.RequestInit;
+import type { RequestInit, Response } from "node-fetch";
+import nodeFetchModuleInjectable from "./unsafe-fetch.injectable";
 
 export type Fetch = (url: string, init?: RequestInit) => Promise<Response>;
 
 const fetchInjectable = getInjectable({
   id: "fetch",
   instantiate: (di): Fetch => {
-    const { httpsProxy, allowUntrustedCAs } = di.inject(userStoreInjectable);
-    const agent = httpsProxy
-      ? new HttpsProxyAgent({
-        proxy: httpsProxy,
-        rejectUnauthorized: !allowUntrustedCAs,
-      })
-      : undefined;
+    const { default: fetch } = di.inject(nodeFetchModuleInjectable);
 
-    return (url, init = {}) => fetch(url, {
-      agent,
-      ...init,
-    });
+    return (url, init) => fetch(url, init);
   },
   causesSideEffects: true,
 });
